+++ filesystem/src/main/java/com/dropbox/android/external/fs3/DiskRead.kt
@@ -1,4 +1,4 @@
-package com.dropbox.android.external.store4
+package com.dropbox.android.external.fs3
 
 /**
  *  Interface for retrieving [Raw] data from disk/persistent sources based on a [key] identifier
