+++ filesystem/src/main/java/com/dropbox/android/external/fs3/FSReader.kt
@@ -1,7 +1,6 @@
 package com.dropbox.android.external.fs3
 
 import com.dropbox.android.external.fs3.filesystem.FileSystem
-import com.dropbox.android.external.store4.DiskRead
 import okio.BufferedSource
 import java.io.FileNotFoundException
