+++ filesystem/src/main/java/com/dropbox/android/external/fs3/Persister.kt
@@ -1,6 +1,5 @@
 package com.dropbox.android.external.fs3
 
-import com.dropbox.android.external.store4.DiskRead
 import com.dropbox.android.external.store4.DiskWrite
 
 /**
