+++ filesystem/src/main/java/com/dropbox/android/external/fs3/SourceFileReader.kt
@@ -2,7 +2,6 @@ package com.dropbox.android.external.fs3
 
 import com.dropbox.android.external.fs3.SourcePersister.Companion.pathForBarcode
 import com.dropbox.android.external.fs3.filesystem.FileSystem
-import com.dropbox.android.external.store4.DiskRead
 import com.dropbox.android.external.store4.legacy.BarCode
 import okio.BufferedSource
 import kotlin.time.Duration
